Add TransferViewModel as parent with handling for settings
